requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
Pillow>=10.0.0

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# GUI Framework
# Note: tkinter comes with Python standard library

# Video and AI processing
opencv-python>=4.8.0
yt-dlp>=2023.7.6
python-dotenv>=1.0.0

# LLM providers (choose one or more)
anthropic>=0.25.0      # For Claude (default)
openai>=1.0.0          # For OpenAI GPT models (optional)
# requests>=2.31.0     # For local LLM support (already included above)

# Data handling
numpy>=1.24.0
urllib3>=2.0.0

# Optional: For enhanced image processing
# scikit-image>=0.21.0

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.7.0
# flake8>=6.0.0